Restructed the FormSetting to separate pages
